Prepare is_it_ok.sh for using in Github Actions tests #4
is_it_ok.sh is a school script for checking final archive. We will
use it for part of basic tests (see issue #4). It needs to be
a little modified, because it doesn't return any exit code etc.

I chose a patch style edition of this file, because its actively
developed by pro project initiator and it could change any time.
With *.patch file its only about uploading new version of the file
to the repository - everything else will be done automatically.